Refactor(src/*): Use src folder to reorganize project
