+++ backend/src/api/controllers/equipamentosController.js
