+++ backend/src/api/models/equipamentoSNModel.js
