+++ backend/src/api/routes/equipamentos.js
