+++ backend/src/api/services/equipamentosService.js
