+++ backend/src/db/equipamentos.json
