+++ backend/src/di/equipamentoInjector.js
