+++ package.json
@@ -1,34 +1,10 @@
 
 {
-    "name": "backend",
+    "name": "reservacin",
     "version": "1.0.0",
-    "description": "Backend com NodeJS para comunicação com banco de dados",
+    "description": "Reservas de salas e equipamentos",
     "main": "app.js",
-    "scripts": {
-        "start": "nodemon app.js",
-        "test": "jest"
-    },
     "author": "chmfc, jrvn, js3, jcasf, mgc3, thrb, tns",
-    "license": "ISC",
-    "dependencies": {
-        "consign": "^0.1.6",
-        "cors": "^2.8.5",
-        "dependencies": "^0.0.1",
-        "dotenv-safe": "^9.1.0",
-        "express": "^4.19.2",
-        "shortid": "^2.2.16",
-        "supertest": "^7.0.0"
-    },
-    "devDependencies": {
-        "@cucumber/cucumber": "^10.7.0",
-        "@types/jest": "^29.5.12",
-        "cucumber": "^6.0.7",
-        "jest": "^29.7.0",
-        "jest-cucumber": "^4.4.0",
-        "jest-haste-map": "^29.7.0",
-        "nodemon": "^3.1.3",
-        "swagger-jsdoc": "^6.2.8",
-        "swagger-ui-express": "^5.0.1"
-    }
+    "license": "ISC"
 }
